LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-reject-expected.txt


FAIL event.intercept() should abort if the given promise rejects assert_equals: expected "#1" but got ""
